Subscription refactoring
